# Don't allocate huge slices to merge pgids in freelist.write