@benbjohnson 2016-12-21
@josharian Thanks for looking at Bolt performance. I can take a look at this pull request more in-depth later today. I don't mind if the synthetic benchmarks are worse so long as real world benchmarks are showing good improvement. 30% reduction is great!