@benbjohnson 2016-12-21
@josharian Do you have a real world benchmark available that I can use to help test?