@josharian 2016-12-21
I'm not entirely sure how/why my db and usage patterns elicit this behavior, but I can try to write something synthetic that reproduces it. I'm off for the holidays so it may take a few days.