@benbjohnson 2016-12-21
@josharian Cool, no rush. I'd like to start collecting a better variety of benchmarks that are closer to real world usage than my synthetic benchmarks. Tuning performance for one person tends to make it worse for others so it'd be nice to have more use cases.