@funny-falcon 2016-12-21
use of copyall in freelist.write is obviously right idea.