@josharian 2016-12-23
I don't have a benchmark written for this yet. Writing a benchmark may be hard, since it requires a very large file. I do, however, understand more about what's going on. I'll file a new issue to discuss.